requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-docx
matplotlib
python-dateutil
requests
